requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.26.0
opencv-python-headless>=4.9.0.80
rapidfuzz>=3.6.0
